+++ src/components/CoreConcept.css
@@ -1,0 +1,19 @@
+#core-concepts li {
+  width: 10rem;
+  text-align: center;
+}
+
+#core-concepts li img {
+  height: 8rem;
+  width: 12rem;
+  object-fit: cover;
+}
+
+#core-concepts li h3 {
+  margin: 0.5rem 0;
+}
+
+#core-concepts li p {
+  font-size: 0.9rem;
+  text-align: left;
+}
