+++ src/components/CoreConcepts.css
@@ -7,23 +7,3 @@
   flex-wrap: wrap;
   justify-content: center;
 }
-
-#core-concepts li {
-  width: 10rem;
-  text-align: center;
-}
-
-#core-concepts li img {
-  height: 8rem;
-  width: 12rem;
-  object-fit: cover;
-}
-
-#core-concepts li h3 {
-  margin: 0.5rem 0;
-}
-
-#core-concepts li p {
-  font-size: 0.9rem;
-  text-align: left;
-}
