+++ src/components/CoreConcepts.jsx
@@ -1,18 +1,16 @@
-import reactImg from "../assets/react.svg";
 import { CORE_CONCEPTS } from "../data";
+import CoreConcept from "./CoreConcept";
 import "./CoreConcepts.css";
-const reactDescriptions = ["Fundamental", "Crucial", "Core"];
+import Section from "./Section";
 
-function CoreConcepts({ alt, description, image, title }) {
+export default function CoreConcepts() {
   return (
-    <div>
-      <li>
-        <img src={image} alt={alt} />
-        <h3>{title}</h3>
-        <p>{description}</p>
-      </li>
-    </div>
+    <Section title="Core Concepts" id="core-concepts">
+      <ul>
+        {CORE_CONCEPTS.map((element) => (
+          <CoreConcept key={element.title} {...element} />
+        ))}
+      </ul>
+    </Section>
   );
 }
-
-export default CoreConcepts;
